requirements.txt
Django
django-rest-framework
mongoengine
django-rest-framework-mongoengine

